feat: add subgraph health indicators
