+++ apps/web/src/pages/swap.tsx
@@ -1,4 +1,7 @@
 import { CHAIN_IDS } from 'utils/wagmi'
+
+import { V3SubgraphHealthIndicator } from 'components/SubgraphHealthIndicator'
+
 import Swap from '../views/Swap'
 import { SwapFeaturesProvider } from '../views/Swap/SwapFeaturesContext'
 
@@ -6,6 +9,7 @@ const SwapPage = () => {
   return (
     <SwapFeaturesProvider>
       <Swap />
+      <V3SubgraphHealthIndicator />
     </SwapFeaturesProvider>
   )
 }
